Misc/message-bot/src/requirements.txt
dnspython==1.16.0
PyNaCl==1.3.0
python-dotenv==0.21.0
discord-ext-bot==1.0.1
discord==2.1.0
Flask-RESTful==0.3.9
google==3.0.0
google-api-python-client==2.70.0
gspread==5.7.2
oauth2client==4.1.3
